requirements.txt
python-dotenv
gitpython
redis
Flask
flask_restful
testresources
requests